+++ SleepModel/HEMSleepPillFinderViewController.h
@@ -8,11 +8,13 @@
 
 #import <UIKit/UIKit.h>
 #import "HEMBaseController.h"
+#import "HEMSleepPillDFUDelegate.h"
 
 @class HEMDeviceService;
 
 @interface HEMSleepPillFinderViewController : HEMBaseController
 
 @property (nonatomic, strong) HEMDeviceService* deviceService;
+@property (nonatomic, weak) id<HEMSleepPillDFUDelegate> delegate;
 
 @end