+++ SleepModel/HEMSystemAlertPresenter.h
@@ -12,6 +12,7 @@
 @class HEMDeviceAlertService;
 @class HEMSystemAlertPresenter;
 @class HEMTimeZoneAlertService;
+@class HEMDeviceService;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -30,7 +31,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (instancetype)initWithNetworkAlertService:(HEMNetworkAlertService*)networkAlertService
                          deviceAlertService:(HEMDeviceAlertService*)deviceAlertService
-                       timeZoneAlertService:(HEMTimeZoneAlertService*)tzAlertService NS_DESIGNATED_INITIALIZER;
+                       timeZoneAlertService:(HEMTimeZoneAlertService*)tzAlertService
+                              deviceService:(HEMDeviceService*)deviceService NS_DESIGNATED_INITIALIZER;
 - (instancetype)init NS_UNAVAILABLE;
 
 - (void)bindWithContainerView:(UIView*)containerView;